refactor metamask fixtures
