+++ bridge-ui/test/utils/index.ts
@@ -1,0 +1,1 @@
+export { selectTokenAndWaitForBalance } from './selectTokenAndWaitForBalance';
